requirements.txt
blinker==1.4
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask-nav==0.6
Flask-Navigation==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
visitor==0.1.3
Werkzeug==1.0.1

mysql-connector-python~=8.0.25